moved group entry attributes funcs
